--- a/core.go
+++ b/core.go
@@ -7,7 +7,6 @@ import (
 	"encoding/binary"
 	"encoding/hex"
 	"fmt"
-	"io"
 	"os"
 )
 
@@ -24,19 +23,12 @@ func (s Score) String() string {
 	return hex.EncodeToString(s[:])
 }
 
-// addr is index's address type alias
-type addr [2]int
-
-// index is index type alias
-type index map[Score]addr
-
 // Store represents a data store.
 type Store struct {
 	name  string
-	idx   index
 	eof   int
 	data  *os.File
-	index *os.File
+	index indexer
 }
 
 // New creates a new empty storage
@@ -67,91 +59,22 @@ func Open(storeName string) (s *Store, err error) {
 	if err != nil {
 		return
 	}
-	idx, err := loadIndex(storeName)
-	if err != nil && !os.IsNotExist(err) {
-		return
-	}
-	indexFileName := fmt.Sprintf("%s.idx", storeName)
-	ifh, err := os.OpenFile(indexFileName, os.O_CREATE|os.O_WRONLY, 0600)
+	idx, err := openIndex(storeName)
 	if err != nil {
 		return
-	}
-	if len(idx) == 0 {
-		idx = buildIndex(dfh, ifh)
 	}
 	s = &Store{
 		name:  storeName,
-		idx:   idx,
 		eof:   int(i.Size()),
 		data:  dfh,
-		index: ifh,
+		index: idx,
 	}
 	return
-}
-
-func loadIndex(storeName string) (idx index, err error) {
-	indexFileName := fmt.Sprintf("%s.idx", storeName)
-	f, err := os.OpenFile(indexFileName, os.O_RDONLY, 0600)
-	if err != nil {
-		return
-	}
-	defer f.Close()
-	idx = make(index)
-	bufSize := 2*IntSize + ScoreSize
-	buf := make([]byte, bufSize, bufSize)
-	for {
-		if _, err = f.Read(buf); err == io.EOF {
-			break
-		}
-		var score Score
-		pos := int(binary.BigEndian.Uint32(buf[0:IntSize]))
-		len := int(binary.BigEndian.Uint32(buf[IntSize : 2*IntSize]))
-		copy(score[:], buf[2*IntSize:])
-		idx[score] = addr{pos, len}
-	}
-	if err == io.EOF {
-		return idx, nil
-	}
-	return
-}
-
-func buildIndex(f *os.File, i *os.File) index {
-	pos := IntSize
-	idx := make(index)
-	lenBuf := make([]byte, IntSize)
-	for {
-		if _, err := f.Read(lenBuf); err == io.EOF {
-			break
-		}
-		len := int(binary.BigEndian.Uint32(lenBuf))
-		buf := make([]byte, len)
-		if _, err := f.Read(buf); err == io.EOF {
-			break
-		}
-		score := makeScore(buf)
-		idx[score] = addr{pos, len}
-		err := writeScore(i, score, pos, len)
-		if err != nil {
-			break
-		}
-		pos += len + IntSize
-	}
-	return idx
 }
 
 func makeScore(b []byte) Score {
 	score := md5.Sum(b)
 	return score
-}
-
-func writeScore(f *os.File, score Score, pos, len int) error {
-	bufSize := 2*IntSize + ScoreSize
-	buf := make([]byte, bufSize, bufSize)
-	binary.BigEndian.PutUint32(buf[0:IntSize], uint32(pos))
-	binary.BigEndian.PutUint32(buf[IntSize:2*IntSize], uint32(len))
-	copy(buf[2*IntSize:], score[:])
-	_, err := f.Write(buf)
-	return err
 }
 
 // Name returns name of a store
@@ -166,12 +89,11 @@ func (s *Store) MakeScore(b []byte) Score {
 
 // Read a data for a given score
 func (s *Store) Read(score Score) (b []byte, err error) {
-	addr, ok := s.idx[score]
+	pos, len, ok := s.index.get(score)
 	if !ok {
 		err = fmt.Errorf("Unknown score %s", score)
 		return
 	}
-	pos, len := addr[0], addr[1]
 	b = make([]byte, len)
 	_, err = s.data.ReadAt(b, int64(pos))
 	if err != nil {
@@ -187,7 +109,7 @@ func (s *Store) Read(score Score) (b []byte, err error) {
 // Write given data and return it's score
 func (s *Store) Write(b []byte) (score Score, err error) {
 	score = s.MakeScore(b)
-	if _, ok := s.idx[score]; ok {
+	if _, _, ok := s.index.get(score); ok {
 		return
 	}
 	len := len(b)
@@ -204,15 +126,17 @@ func (s *Store) Write(b []byte) (score Score, err error) {
 	}
 	newPos := s.eof + IntSize
 	newLen := n - IntSize
-	s.idx[score] = addr{newPos, newLen}
-	writeScore(s.index, score, newPos, newLen)
+	err = s.index.put(score, newPos, newLen)
+	if err != nil {
+		return
+	}
 	s.eof += n
 	return
 }
 
 // Close provided storage
 func (s *Store) Close() error {
-	err := s.index.Close()
+	err := s.index.close()
 	if err != nil {
 		return err
 	}
@@ -221,11 +145,7 @@ func (s *Store) Close() error {
 
 // Delete provided storage
 func (s *Store) Delete() error {
-	if err := s.Close(); err != nil {
-		return err
-	}
-	indexFileName := fmt.Sprintf("%s.idx", s.name)
-	if err := os.Remove(indexFileName); err != nil {
+	if err := s.index.delete(); err != nil {
 		return err
 	}
 	dataFileName := fmt.Sprintf("%s.data", s.name)