--- a/core.go
+++ b/core.go
@@ -4,7 +4,6 @@ package beansdb
 import (
 	"crypto/md5"
 	"crypto/rand"
-	"encoding/binary"
 	"encoding/hex"
 	"fmt"
 	"os"
@@ -26,8 +25,7 @@ func (s Score) String() string {
 // Store represents a data store.
 type Store struct {
 	name  string
-	eof   int
-	data  *os.File
+	data  datalogger
 	index indexer
 }
 
@@ -49,24 +47,18 @@ func New() (s *Store, err error) {
 }
 
 // Open opens provided storage
-func Open(storeName string) (s *Store, err error) {
-	dataFileName := fmt.Sprintf("%s.data", storeName)
-	dfh, err := os.OpenFile(dataFileName, os.O_APPEND|os.O_RDWR, 0600)
+func Open(name string) (s *Store, err error) {
+	data, err := openDataLog(name)
 	if err != nil {
 		return
 	}
-	i, err := dfh.Stat()
-	if err != nil {
-		return
-	}
-	idx, err := openIndex(storeName)
+	idx, err := openIndex(name)
 	if err != nil {
 		return
 	}
 	s = &Store{
-		name:  storeName,
-		eof:   int(i.Size()),
-		data:  dfh,
+		name:  name,
+		data:  data,
 		index: idx,
 	}
 	return
@@ -89,18 +81,14 @@ func (s *Store) MakeScore(b []byte) Score {
 
 // Read a data for a given score
 func (s *Store) Read(score Score) (b []byte, err error) {
-	pos, len, ok := s.index.get(score)
+	p, l, ok := s.index.get(score)
 	if !ok {
 		err = fmt.Errorf("Unknown score %s", score)
 		return
 	}
-	b = make([]byte, len)
-	_, err = s.data.ReadAt(b, int64(pos))
-	if err != nil {
-		return
-	}
+	b, err = s.data.get(p, l)
 	if score != s.MakeScore(b) {
-		b = []byte{}
+		b = nil
 		err = fmt.Errorf("Checksum failure")
 	}
 	return
@@ -112,35 +100,23 @@ func (s *Store) Write(b []byte) (score Score, err error) {
 	if _, _, ok := s.index.get(score); ok {
 		return
 	}
-	len := len(b)
-	buf := make([]byte, len+IntSize, len+IntSize)
-	binary.BigEndian.PutUint32(buf[:IntSize], uint32(len))
-	copy(buf[IntSize:], b)
-	n, err := s.data.Write(buf)
+	p, l, err := s.data.put(b)
 	if err != nil {
 		return
 	}
-	err = s.data.Sync()
+	err = s.index.put(score, p, l)
 	if err != nil {
 		return
 	}
-	newPos := s.eof + IntSize
-	newLen := n - IntSize
-	err = s.index.put(score, newPos, newLen)
-	if err != nil {
-		return
-	}
-	s.eof += n
 	return
 }
 
 // Close provided storage
 func (s *Store) Close() error {
-	err := s.index.close()
-	if err != nil {
+	if err := s.index.close(); err != nil {
 		return err
 	}
-	return s.data.Close()
+	return s.data.close()
 }
 
 // Delete provided storage
@@ -148,6 +124,5 @@ func (s *Store) Delete() error {
 	if err := s.index.delete(); err != nil {
 		return err
 	}
-	dataFileName := fmt.Sprintf("%s.data", s.name)
-	return os.Remove(dataFileName)
+	return s.data.delete()
 }