… testdata/fox-dog.data
… testdata/fox-dog.data.golden
… testdata/fox-dog.idx.golden