Doesn't seems to break anything much

```
benchmark            old ns/op     new ns/op     delta
BenchmarkOpen-4      508112997     505983031     -0.42%
BenchmarkWrite-4     110976        122633        +10.50%
BenchmarkRead-4      1993          2156          +8.18%

benchmark            old allocs     new allocs     delta
BenchmarkOpen-4      9582           9633           +0.53%
BenchmarkWrite-4     2              2              +0.00%
BenchmarkRead-4      1              1              +0.00%

benchmark            old bytes     new bytes     delta
BenchmarkOpen-4      41046620      41070212      +0.06%
BenchmarkWrite-4     1312          1312          +0.00%
BenchmarkRead-4      16            16            +0.00%
```